fix:修正首頁Section3裝飾線條
